# docs/swagger.yaml
basePath: /v1
definitions:
...
  errs.Error:
    properties:
      caller:
        type: string
      details:
        items:
          $ref: '#/definitions/errs.FieldError'
        type: array
      message:
        type: string
      reason:
        type: string
      type:
        $ref: '#/definitions/errs.Type'
    type: object
  errs.FieldError:
    properties:
      field:
        type: string
      message:
        type: string
      rule:
        type: string
    type: object
  errs.Type:
    enum:
    - VALIDATION
    - NOT_FOUND
    - CONFLICT
    - TIMEOUT
    - INTERNAL
    type: string
    x-enum-comments:
      ErrTypeConflict: "409"
      ErrTypeInternal: "500"
      ErrTypeNotFound: "404"
      ErrTypeTimeout: "504"
      ErrTypeValidation: "422"
    x-enum-descriptions:
    - "422"
    - "404"
    - "409"
    - "504"
    - "500"
    x-enum-varnames:
    - ErrTypeValidation
    - ErrTypeNotFound
    - ErrTypeConflict
    - ErrTypeTimeout
    - ErrTypeInternal
...
  out.DocumentListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/out.DocumentResponse'
        type: array
      total:
        type: integer
    type: object
  out.DocumentResponse:
    properties:
      chunkCount:
        type: integer
      createdAt:
        type: string
      domain:
        example: flotation
        type: string
      id:
        type: string
      language:
        example: ru
        type: string
      sourceType:
        example: book
        type: string
      title:
        type: string
    type: object
...
  out.IngestResponse:
    properties:
      chunksIngested:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 'Evidence-Backed Hypothesis Factory: генерация проверяемых технологических
    гипотез по обогащению руд на основе базы знаний (RAG).'
  title: Фабрика гипотез — Hypothesis Factory API
  version: "1.0"
paths:
  /documents:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/out.DocumentListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.Error'
      summary: Список документов базы знаний
      tags:
      - documents
    post:
      consumes:
      - multipart/form-data
      description: Парсит PDF/DOCX/XLSX/изображение через Docling, эмбеддит чанки
        BGE-M3, пишет в pgvector.
      parameters:
      - description: Файл документа
        in: formData
        name: file
        required: true
        type: file
      - description: Заголовок (по умолчанию — имя файла)
        in: formData
        name: title
        type: string
      - default: report
        description: book | regulation | scheme | historical_case | report | article
          | patent (article = GROBID + Semantic Scholar; patent — через обычный Docling-путь,
          structured-парсинг формулы изобретения не делаем — используйте authors/year
          для инвентора/даты подачи)
        in: formData
        name: sourceType
        type: string
      - default: flotation
        description: Домен знаний
        in: formData
        name: domain
        type: string
      - default: ru
        description: Язык
        in: formData
        name: language
        type: string
      - description: Авторы (через запятую), если известны
        in: formData
        name: authors
        type: string
      - description: Год издания/публикации, если известен
        in: formData
        name: year
        type: string
      - description: Издание/версия, если известно
        in: formData
        name: edition
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/out.IngestResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.Error'
      summary: Загрузка документа в базу знаний
      tags:
      - documents
  /documents/{documentId}:
    delete:
      parameters:
      - description: UUID документа
        in: path
        name: documentId
        required: true
        type: string
      responses:
        "204":
          description: Документ удалён
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errs.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.Error'
      summary: Удаление документа из базы знаний
      tags:
      - documents
...
swagger: "2.0"